extract AssetConditionEvaluator
