+++ python_modules/dagster/dagster/_core/definitions/asset_daemon_cursor.py
@@ -67,6 +67,8 @@ class ObserveRequestTimestampSerializer(FieldSerializer):
     }
 )
 @dataclass(frozen=True)
+# TODO: rename to scheduling cursor or something
+# 2024-05-16 -- schrockn
 class AssetDaemonCursor:
     """State that's stored between daemon evaluations.
